Implement originProv
Add a boolean isOrigin to reflect whether a provenance is an originProv or not.
When printing a provenance there is "o:" in front of an originProv.
Print a note explaining where the type is defined when there is an originProv.
Filter provenances so that the one printed for the error is not an originProv.