Generate let bindings for class parameters during desugaring
We used to generate let bindings which destruct class parameters
during normalization. This works fine but it exposes some
inflexibility. For example, in the previous commit, I added a case
where free variables from a absorbed case branch are shadowed.
If we moving let bindings generation to desugaring stage, the
problem can be solved easily. This change also paves the road for
duplicated branch lifting.