Improved specialisation
Improved specialisation by making spec points be application sites
rather than individual parameters. This allows for the flows of types to
be tracked neater through a program and allow for less unnecessary code
to be generated.

In addition added better suppoer for higher level specialisation —
passing specialised functions to other functions — however this is still
buggy.